Take `&'static CStr` in `FontDescription::set_family_static`
This commit fixes a memory safety issue in this method.

`<str as ToGlibPtr>::to_glib_none` returns a null-terminated string
which only lives as long as the `Stash` it returns. This leads to
use-after-free because `pango_font_description_set_family_static`
expects the given string to outlive this `PangoFontDescription`.

Obviously, we can't simply pass the value of `str::as_ptr` because it
might not be null-terminated. What we need is `&'static CStr` - `CStr`
guarantees the memory representation of the string is null-terminated,
and `&'static` guarantees it will remain valid throughout the program's
lifetime.